requirements.txt
cryptography
python-bitcoinlib
